backend/requirements.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
channels==4.3.2
channels-redis==4.3.0
python-decouple==3.8
psycopg2-binary==2.9.10
redis==7.4.0
google-auth==2.49.1
requests==2.32.5
pillow==12.1.1
asgiref==3.11.1
sqlparse==0.5.5
PyJWT==2.12.0
cryptography==46.0.5
certifi==2026.2.25
charset-normalizer==3.4.5
idna==3.11
urllib3==2.6.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
cffi==2.0.0
pycparser==3.0
tzdata==2025.3
gunicorn==23.0.0
